Rename, reduce redundancy, dont load table if still loading
