--- src/views/Project/bugs.graphql
@@ -1,21 +1,0 @@
-query Bugs($search: BugSearch!, $paging: Paging) {
-  bugs(search: $search, paging: $paging) {
-    edges {
-      node {
-        assignedTo {
-          name
-        }
-        component
-        lastChanged
-        id
-        keywords
-        tags
-        summary
-        status
-        resolution
-        url
-        whiteboard
-      }
-    }
-  }
-}
