+++ src/views/Project/index.jsx
@@ -20,8 +20,8 @@ import AppBar from '../../components/AppBar';
 import ErrorPanel from '../../components/ErrorPanel';
 import TasksTable from '../../components/TasksTable';
 import issuesQuery from './issues.graphql';
-import bugsQuery from './bugs.graphql';
-import commentsQuery from './comments.graphql';
+import bugsQuery from '../bugs.graphql';
+import commentsQuery from '../comments.graphql';
 import {
   GOOD_FIRST_BUG,
   BUGZILLA_STATUSES,
