+++ src/views/bugs.graphql
